Refactoring knapsack module to be a package
