+++ geco/mips/knapsack/yang.py
@@ -1,7 +1,8 @@
 import math
 
-import pyscipopt as scip
 from networkx.utils import py_random_state
+
+from geco.mips.knapsack.generic import knapsack
 
 
 @py_random_state(1)
@@ -26,23 +27,3 @@ def yang_parameter(n, seed=0):
     capacity = math.floor(sum(weights) / 5)
 
     return profits, weights, capacity
-
-
-def knapsack(weights, profits, capacity, name="Knapsack"):
-    assert len(weights) == len(profits)
-    assert capacity >= 0
-    assert all(w >= 0 for w in weights)
-    assert all(p >= 0 for p in profits)
-
-    model = scip.Model(name)
-    # add variables and their cost
-    variables = [model.addVar(lb=0, ub=1, obj=profit, vtype="B") for profit in profits]
-    # add constraints
-    model.addCons(
-        scip.quicksum(weight * variable for weight, variable in zip(weights, variables))
-        <= capacity
-    )
-
-    model.setMaximize()
-
-    return model
